Address comments from Wenchen
